requirements.txt
discord.py==2.3.2
apscheduler==3.10.4
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pytz==2023.3
python-dotenv==1.0.0
